cli: rename status command to info
