+++ pkg/cli/cli.go
@@ -16,8 +16,8 @@
 package cli
 
 import (
+	"github.com/fuddle-io/fuddle/pkg/cli/info"
 	"github.com/fuddle-io/fuddle/pkg/cli/start"
-	"github.com/fuddle-io/fuddle/pkg/cli/status"
 	"github.com/spf13/cobra"
 )
 
@@ -37,7 +37,7 @@ func init() {
 
 	fuddleCmd.AddCommand(
 		start.Command,
-		status.Command,
+		info.Command,
 		demoCmd,
 		versionCmd,
 	)
