+++ pkg/cli/info/command.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package status
+package info
 
 import (
 	"context"
@@ -27,7 +27,7 @@ import (
 )
 
 var Command = &cobra.Command{
-	Use:   "status",
+	Use:   "info",
 	Short: "inspect the status of the cluster",
 }
