+++ pkg/cli/info/flags.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package status
+package info
 
 var (
 	// addr is the Fuddle registry server to query.
